Initial commit of Sunday Reader Typica
This was written to be compiled in Linux, but I don't have a Linux
install on hand with which to test it at the moment.  It is guaranteed
to not compile is it sits, because the texmf folder is intended to be in
ones home directory.  Other than that, the document requires
lilypond/lilypond-book, pdflatex, pdfbook, sed, and multiple latex
packages including titlesec, hyperef, and vectorian.  That vectorian
package is likely the harding to install; I remember having to install
it manually.